Refactor spelling, imports and graddle
