+++ android/project.properties
@@ -6,4 +6,4 @@
 # proguard.config=${sdk.dir}/tools/proguard/proguard-android.txt:proguard-rules.pro
 
 # Project target.
-target=android-19
+# target=android-19
