+++ core/src/com/geometric/wars/cube/DynamicCubeBuilder.java
@@ -6,8 +6,8 @@ import com.badlogic.gdx.graphics.g3d.utils.ModelBuilder;
 import com.badlogic.gdx.math.Matrix4;
 import com.badlogic.gdx.math.Quaternion;
 import com.badlogic.gdx.math.Vector3;
-import com.geometric.wars.Direction3D;
-import com.geometric.wars.Values;
+import com.geometric.wars.utils.Direction3D;
+import com.geometric.wars.utils.Values;
 import com.geometric.wars.cube.mountables.Mountable;
 import com.geometric.wars.models.BoxBuilder;
 import com.geometric.wars.models.DynamicCubeModelDisposer;
