+++ core/src/com/geometric/wars/cube/DynamicCubeView.java
@@ -3,8 +3,8 @@ package com.geometric.wars.cube;
 
 import com.badlogic.gdx.graphics.g3d.Model;
 import com.badlogic.gdx.graphics.g3d.ModelInstance;
-import com.geometric.wars.Direction2D;
-import com.geometric.wars.Values;
+import com.geometric.wars.utils.Direction2D;
+import com.geometric.wars.utils.Values;
 import com.geometric.wars.math.RotationCalculator;
 
 /**
