+++ core/src/com/geometric/wars/enviromentparts/Wall.java
@@ -5,7 +5,7 @@ import com.badlogic.gdx.graphics.g3d.ModelBatch;
 import com.badlogic.gdx.graphics.g3d.ModelCache;
 import com.badlogic.gdx.graphics.g3d.ModelInstance;
 import com.badlogic.gdx.math.Vector3;
-import com.geometric.wars.Values;
+import com.geometric.wars.utils.Values;
 import com.geometric.wars.gameobjects.StaticGameObject;
 import com.geometric.wars.models.WallModel;
