+++ core/src/com/geometric/wars/models/LineModel.java
@@ -5,7 +5,7 @@ import com.badlogic.gdx.graphics.g3d.Material;
 import com.badlogic.gdx.graphics.g3d.Model;
 import com.badlogic.gdx.graphics.g3d.utils.MeshPartBuilder;
 import com.badlogic.gdx.graphics.g3d.utils.ModelBuilder;
-import com.geometric.wars.Values;
+import com.geometric.wars.utils.Values;
 
 /**
  * Class contains instance of Line model.
