+++ core/src/com/geometric/wars/models/WallModel.java
@@ -6,7 +6,7 @@ import com.badlogic.gdx.graphics.g3d.Material;
 import com.badlogic.gdx.graphics.g3d.Model;
 import com.badlogic.gdx.graphics.g3d.attributes.ColorAttribute;
 import com.badlogic.gdx.graphics.g3d.utils.ModelBuilder;
-import com.geometric.wars.Values;
+import com.geometric.wars.utils.Values;
 
 /**
  * Class contains instance of Wall model.
