+++ core/src/com/geometric/wars/player/bots/randomactingbot/RandomBotCube.java
@@ -1,9 +1,7 @@
 package com.geometric.wars.player.bots.randomactingbot;
 
 import com.badlogic.gdx.graphics.Color;
-import com.geometric.wars.Direction3D;
-import com.geometric.wars.cube.DynamicCube;
-import com.geometric.wars.cube.DynamicCubeView;
+import com.geometric.wars.utils.Direction3D;
 import com.geometric.wars.cube.DynamicCubeBuilder;
 import com.geometric.wars.cube.mountables.guns.SimpleGun;
 import com.geometric.wars.player.PlayersCube;
