+++ core/src/com/geometric/wars/player/person/PersonsCube.java
@@ -1,10 +1,10 @@
 package com.geometric.wars.player.person;
 
 import com.badlogic.gdx.graphics.Color;
-import com.geometric.wars.Direction3D;
+import com.geometric.wars.utils.Direction3D;
+import com.geometric.wars.cube.DynamicCubeBuilder;
 import com.geometric.wars.cube.mountables.guns.SimpleGun;
 import com.geometric.wars.input.InputController;
-import com.geometric.wars.cube.DynamicCubeBuilder;
 import com.geometric.wars.player.PlayersCube;
 
 public class PersonsCube extends PlayersCube {
